Neo4j Restructure
