+++ aossl/neo4j/include/neo4j_interface.h
@@ -28,8 +28,39 @@ THE SOFTWARE.
 #include <string>
 #include <unordered_map>
 #include <exception>
+#include <neo4j-client.h>
 
 #include "aossl/core/include/slot_pool.h"
+
+// Neo4j Exception
+
+// A child class of std::exception
+// which holds error information
+struct Neo4jException: public std::exception {
+  // An error message passed on initialization
+  std::string int_msg;
+  const char * int_msg_cstr;
+  neo4j_result_stream_t *results = NULL;
+
+  // Create a Neo4j Exception, and store the given error message
+  inline Neo4jException(std::string msg) {
+    int_msg = "Error in Neo4j Request: " + msg;
+    int_msg_cstr = int_msg.c_str();
+  }
+
+  inline Neo4jException(std::string msg, neo4j_result_stream_t *r) {
+    int_msg = "Error in Neo4j Request: " + msg;
+    int_msg_cstr = int_msg.c_str();
+    results = r;
+  }
+
+  Neo4jException() {}
+  ~Neo4jException() throw() {if (results) {neo4j_close_results(results);}}
+  // Show the error message in readable format
+  const char * what() const throw() {
+    return int_msg_cstr;
+  }
+};
 
 // Query Results
