+++ aossl/neo4j/makefile
@@ -3,7 +3,7 @@ CC = g++
 CFLAGS  = -g -Wall
 STD = -std=c++11
 INCL_DIRS = -I$(CURDIR)/../../
-INCL = $(INSTALL_DIR)/neo4j/include/factory_neo4j.h	$(INSTALL_DIR)/neo4j/include/neo4j_admin.h $(INSTALL_DIR)/neo4j/include/neo4j_interface.h
+INCL = $(INSTALL_DIR)/neo4j/include/factory_neo4j.h	$(INSTALL_DIR)/neo4j/include/neo4j_admin.h $(INSTALL_DIR)/neo4j/include/neo4j_interface.h $(INSTALL_DIR)/neo4j/include/neo4j_connection_pool.h $(INSTALL_DIR)/neo4j/include/neo4j_list.h $(INSTALL_DIR)/neo4j/include/neo4j_map.h $(INSTALL_DIR)/neo4j/include/neo4j_object.h $(INSTALL_DIR)/neo4j/include/neo4j_query_parameter.h $(INSTALL_DIR)/neo4j/include/neo4j_result_iterator.h $(INSTALL_DIR)/neo4j/include/neo4j_result_tree.h $(INSTALL_DIR)/neo4j/include/neo4j_utils.h
 INCL_DIR = $(INSTALL_DIR)/neo4j
 LIBS = -lneo4j-client -lssl -lcrypto -lm
 RM = rm -f
@@ -35,6 +35,30 @@ $(INSTALL_DIR)/neo4j/include/neo4j_admin.h: include/neo4j_admin.h
 	cp $< $@
 
 $(INSTALL_DIR)/neo4j/include/neo4j_interface.h: include/neo4j_interface.h
+	cp $< $@
+
+$(INSTALL_DIR)/neo4j/include/neo4j_connection_pool.h: include/neo4j_connection_pool.h
+	cp $< $@
+
+$(INSTALL_DIR)/neo4j/include/neo4j_list.h: include/neo4j_list.h
+	cp $< $@
+
+$(INSTALL_DIR)/neo4j/include/neo4j_map.h: include/neo4j_map.h
+	cp $< $@
+
+$(INSTALL_DIR)/neo4j/include/neo4j_object.h: include/neo4j_object.h
+	cp $< $@
+
+$(INSTALL_DIR)/neo4j/include/neo4j_query_parameter.h: include/neo4j_query_parameter.h
+	cp $< $@
+
+$(INSTALL_DIR)/neo4j/include/neo4j_result_iterator.h: include/neo4j_result_iterator.h
+	cp $< $@
+
+$(INSTALL_DIR)/neo4j/include/neo4j_result_tree.h: include/neo4j_result_tree.h
+	cp $< $@
+
+$(INSTALL_DIR)/neo4j/include/neo4j_utils.h: include/neo4j_utils.h
 	cp $< $@
 
 # Test
