+++ aossl/neo4j/neo4j_admin.cpp
@@ -24,507 +24,8 @@ THE SOFTWARE.
 
 #include "include/neo4j_admin.h"
 
-// Results Iterator
-
-// Get the next result from the iterator
-ResultTreeInterface* ResultsIterator::next() {
-  neo4j_result_t *res = neo4j_fetch_next(results);
-  if (!res) {return NULL;} else {return new ResultTree(res);}
-}
-
-void ResultsIterator::clear_results() {
-  if (neo4j_close_results(results) != 0) {
-    const char * err_string = strerror(errno);
-    std::string err_msg(err_string);
-    err_msg = "Error clearing results from memory: " + err_msg;
-    throw Neo4jException(err_msg);
-  }
-}
-
-// Result Tree
-
-// Get a DB Object
-DbObjectInterface* ResultTree::get(int index) {
-  return new DbObject(result, index);
-}
-
-// DB Map
-
-// Constructor
-DbMap::DbMap(neo4j_result_t *r, int ind, std::vector<std::string> key_list, \
-  ValueGenerationFunction mf, std::vector<KeyGenerationFunction> mfs) {
-  result = r;
-  index = ind;
-  map_function = mf;
-  for (std::size_t i = 0; i < mfs.size(); i++) {
-    map_functions.push_back(mfs[i]);
-  }
-  for (std::size_t j = 0; j < key_list.size(); j++) {
-    keys.push_back(key_list[j]);
-  }
-}
-
-DbMap::DbMap(neo4j_result_t *r, int ind, std::vector<std::string> key_list, \
-  ValueGenerationFunction mf, std::vector<KeyGenerationFunction> mfs, \
-  int pindex) {
-  result = r;
-  index = ind;
-  map_function = mf;
-  for (std::size_t i = 0; i < mfs.size(); i++) {
-    map_functions.push_back(mfs[i]);
-  }
-  for (std::size_t j = 0; j < key_list.size(); j++) {
-    keys.push_back(key_list[j]);
-  }
-  path_index = pindex;
-}
-
-// Get the value for this node/edge
-void DbMap::get_map() {
-  db_obj = neo4j_null;
-  map = neo4j_null;
-  if (result) {
-    // Get the query result field
-    db_obj = neo4j_result_field(result, index);
-    if ( neo4j_instanceof(db_obj, NEO4J_PATH) ) {
-      // If the DB Value is a path, then we need to update it to the
-      // respective element based on the path index
-      if (path_index != -1) {
-        // Check if we have a node or relationship
-        if (path_index == 0) {
-          // We have a node
-          db_obj = neo4j_path_get_node(db_obj, path_index);
-        } else if (path_index == 1) {
-          // We have a relationship
-          db_obj = neo4j_path_get_relationship(db_obj, 0, NULL);
-        } else if (path_index % 2 == 0) {
-          // We have a node
-          db_obj = neo4j_path_get_node(db_obj, path_index/2);
-        } else {
-          // We have a relationship
-          db_obj = neo4j_path_get_relationship(db_obj, (path_index-1)/2, NULL);
-        }
-      }
-    }
-    if ( neo4j_instanceof(db_obj, NEO4J_NODE) || neo4j_instanceof(db_obj, \
-      NEO4J_RELATIONSHIP) ) {
-      // Get to the Neo4j Map via the map functions
-      map = (*(map_function))(db_obj);
-      for (std::size_t i = 0; i < map_functions.size(); i++) {
-        neo4j_value_t key_str = neo4j_string(keys[i].c_str());
-        map = (*(map_functions[i]))(map, key_str);
-      }
-    }
-  }
-}
-
-bool DbMap::element_exists(std::string key) {
-  std::lock_guard<std::mutex> lock(data_mutex);
-  // Get the value from the map
-  get_map_value(key);
-  if (neo4j_eq(neo4j_null, map_elt)) {
-    return false;
-  }
-  return true;
-}
-
-// Get a value out of the map
-void DbMap::get_map_value(std::string key) {
-  get_map();
-  map_elt = neo4j_map_get(map, key.c_str());
-}
-
-// Get the number of elements in the map
-unsigned int DbMap::size() {
-  std::lock_guard<std::mutex> lock(data_mutex);
-
-  // Retrieve the map
-  get_map();
-
-  // Return the map size
-  return neo4j_map_size(map);
-}
-
-// Get an element out of the map
-
-bool DbMap::get_bool_element(std::string key) {
-  std::lock_guard<std::mutex> lock(data_mutex);
-
-  // Get the value from the map
-  get_map_value(key);
-
-  // Convert the value to a boolean
-  return neo4j_bool_value(map_elt);
-}
-
-int DbMap::get_int_element(std::string key) {
-  std::lock_guard<std::mutex> lock(data_mutex);
-
-  // Get the value from the map
-  get_map_value(key);
-
-  // Convert the value to an int
-  return neo4j_int_value(map_elt);
-}
-
-double DbMap::get_float_element(std::string key) {
-  std::lock_guard<std::mutex> lock(data_mutex);
-
-  // Get the value from the map
-  get_map_value(key);
-
-  // Convert the value to a float
-  return neo4j_float_value(map_elt);
-}
-
-std::string DbMap::get_string_element(std::string key, int char_buffer_size) {
-  std::lock_guard<std::mutex> lock(data_mutex);
-
-  // Get the value from the map
-  get_map_value(key);
-
-  // Convert the value to a string
-  neo4j_string_value(map_elt, db_cstr, char_buffer_size);
-  if (db_cstr && db_cstr[0] != '\0') {
-    elt.assign(db_cstr);
-  } else {elt = "";}
-  return elt;
-}
-
-DbListInterface* DbMap::get_list_element(std::string key) {
-  return new DbList(result, index, map_function, path_index, \
-    neo4j_map_kget, key);
-}
-
-std::string DbMap::to_string() {
-  std::lock_guard<std::mutex> lock(data_mutex);
-  get_map();
-  if (neo4j_eq(neo4j_null, map)) {
-    return "";
-  }
-  neo4j_tostring(map, buf, 128);
-  str_val.assign(buf);
-  return str_val;
-}
-
-// DB List
-
-// Get the value for this node/edge
-neo4j_value_t DbList::get_list() {
-  if (!result) {
-    return neo4j_null;
-  } else {
-    // Get the DB Value
-    neo4j_value_t db_obj = neo4j_result_field(result, index);
-    // If the DB Value is a path, then we need to update it to the
-    // respective element based on the path index
-    if (path_index != -1) {
-      // Check if we have a node or relationship
-      if (path_index == 0) {
-        // We have a node
-        db_obj = neo4j_path_get_node(db_obj, path_index);
-      } else if (path_index == 1) {
-        // We have a relationship
-        db_obj = neo4j_path_get_relationship(db_obj, 0, NULL);
-      } else if (path_index % 2 == 0) {
-        // We have a node
-        db_obj = neo4j_path_get_node(db_obj, path_index/2);
-      } else {
-        // We have a relationship
-        db_obj = neo4j_path_get_relationship(db_obj, (path_index-1)/2, NULL);
-      }
-    }
-    neo4j_value_t list = (*(list_function))(db_obj);
-    if (map_function) {
-      return (*(map_function))(list, neo4j_string(map_key.c_str()));
-    }
-    return list;
-  }
-}
-
-// Get a value from this list
-neo4j_value_t DbList::get_list_value(unsigned int ind) {
-  neo4j_value_t list = get_list();
-  return neo4j_list_get(list, ind);
-}
-
-// Get the number of elements in the list
-unsigned int DbList::size() {
-  // Retrieve the DB Object
-  neo4j_value_t db_obj = get_list();
-
-  // Return the map size
-  return neo4j_list_length(db_obj);
-}
-
-// Get an element out of the list
-
-bool DbList::get_bool_element(unsigned int ind) {
-  // Get the value from the map
-  neo4j_value_t element = get_list_value(ind);
-
-  // Convert the value to a float
-  return neo4j_bool_value(element);
-}
-
-int DbList::get_int_element(unsigned int ind) {
-  // Get the value from the map
-  neo4j_value_t element = get_list_value(ind);
-
-  // Convert the value to a float
-  return neo4j_int_value(element);
-}
-
-double DbList::get_float_element(unsigned int ind) {
-  // Get the value from the map
-  neo4j_value_t element = get_list_value(ind);
-
-  // Convert the value to a float
-  return neo4j_float_value(element);
-}
-
-std::string DbList::get_string_element(unsigned int ind, int char_buffer_size) {
-  // Get the value from the list
-  neo4j_value_t element = get_list_value(ind);
-
-  // Convert the value to a string
-  neo4j_string_value(element, db_cstr, char_buffer_size);
-  elt.assign(db_cstr);
-  return elt;
-}
-
-std::string DbList::to_string() {
-  neo4j_value_t value = get_list();
-  if (neo4j_eq(neo4j_null, value)) {
-    return "";
-  }
-  neo4j_tostring(value, buf, 128);
-  str_val.assign(buf);
-  return str_val;
-}
-
-// DB Object
-
-// Internal
-
-// Get a Value from a Result
-neo4j_value_t DbObject::get_value() {
-  if (!result) {
-    return neo4j_null;
-  } else {
-    val = neo4j_result_field(result, index);
-    // Check if this is the result of getting a value from a path
-    if (path_index != -1) {
-      // Check if we have a node or relationship
-      if (path_index == 0) {
-        // We have a node
-        return neo4j_path_get_node(val, path_index);
-      } else if (path_index == 1) {
-        // We have a relationship
-        return neo4j_path_get_relationship(val, 0, NULL);
-      } else if (path_index % 2 == 0) {
-        // We have a node
-        return neo4j_path_get_node(val, path_index / 2);
-      } else {
-        // We have a relationship
-        return neo4j_path_get_relationship(val, (path_index-1)/2, NULL);
-      }
-    }
-    return val;
-  }
-}
-
-// Determine if a Value is of the specified type
-bool DbObject::is_instance_of(neo4j_type_t type) {
-  neo4j_value_t value = get_value();
-  if (neo4j_eq(neo4j_null, value)) {
-    return false;
-  }
-  if (neo4j_instanceof(value, type)) {
-    return true;
-  }
-  return false;
-}
-
-// General
-
-// Convert a Value to a string representation
-std::string DbObject::to_string() {
-  neo4j_value_t value = get_value();
-  if (neo4j_eq(neo4j_null, value)) {
-    return "";
-  }
-  neo4j_tostring(value, buf, 128);
-  str_val.assign(buf);
-  return str_val;
-}
-
-// Get a map of the properties
-DbMapInterface* DbObject::properties() {
-  if ( is_node() ) {
-    return new DbMap(result, index, neo4j_node_properties, path_index);
-  } else if ( is_edge() ) {
-    return new DbMap(result, index, neo4j_relationship_properties, path_index);
-  } else {
-    std::string err_msg = \
-      "Tried to retrieve properties for invalid object";
-    throw Neo4jException(err_msg);
-  }
-}
-
-// Get a list of node labels
-DbListInterface* DbObject::labels() {
-  if ( is_node() ) {
-    return new DbList(result, index, neo4j_node_labels, path_index);
-  } else {
-    std::string err_msg = \
-      "Tried to retrieve labels for object that is not a node";
-    throw Neo4jException(err_msg);
-  }
-}
-
-// Get the relationship property
-std::string DbObject::type() {
-  neo4j_value_t val = get_value();
-  if ( is_edge() ) {
-    neo4j_value_t rel_type = neo4j_relationship_type(val);
-    char buf[256];
-    neo4j_string_value(rel_type, buf, 256);
-    std::string ret_val(buf);
-    return ret_val;
-  } else {
-    std::string err_msg = \
-      "Tried to retrieve type for object that is not an edge";
-    throw Neo4jException(err_msg);
-  }
-}
-
-bool DbObject::forward() {
-  neo4j_value_t val = get_value();
-  if ( is_edge() && path_index != -1 ) {
-    neo4j_path_get_relationship(val, path_index, &is_forward);
-    return is_forward;
-  } else {
-    std::string err_msg = "Tried to pull relationship dir on invalid object";
-    throw Neo4jException(err_msg);
-  }
-}
-
-neo4j_value_t DbObject::get_identity() {
-  neo4j_value_t val = get_value();
-  if ( is_node() ) {
-    return neo4j_node_identity(val);
-  } else if ( is_edge() ) {
-    return neo4j_relationship_identity(val);
-  } else {
-    std::string err_msg = "Tried to pull properties on invalid object";
-    throw Neo4jException(err_msg);
-  }
-}
-
-unsigned int DbObject::size() {
-  neo4j_value_t val = get_value();
-  if ( is_path() ) {
-    return (neo4j_path_length(val) * 2) + 1;
-  } else {
-    std::string err_msg = "Tried to pull path length on invalid object";
-    throw Neo4jException(err_msg);
-  }
-}
-
-// Neo4j Query Parameters
-
-void Neo4jQueryParameter::add_value(bool new_val) {
-  std::string err_msg;
-  if (!is_list) {
-    err_msg = "Trying to add value to non-arry parameter";
-    throw Neo4jException(err_msg);
-  } else {
-    if (type == -1 || type == _BOOL_TYPE) {
-      bool_values.push_back(new_val);
-      type = _BOOL_TYPE;
-    } else {
-      err_msg = "Only single-type arrays are supported";
-      throw Neo4jException(err_msg);
-    }
-  }
-}
-
-void Neo4jQueryParameter::add_value(std::string new_val) {
-  std::string err_msg;
-  if (!is_list) {
-    err_msg = "Trying to add value to non-arry parameter";
-    throw Neo4jException(err_msg);
-  } else {
-    if (type == -1 || type == _STR_TYPE) {
-      str_values.push_back(new_val);
-      type = _STR_TYPE;
-    } else {
-      err_msg = "Only single-type arrays are supported";
-      throw Neo4jException(err_msg);
-    }
-  }
-}
-
-void Neo4jQueryParameter::add_value(const char * new_val) {
-  std::string new_str(new_val);
-  add_value(new_str);
-}
-
-void Neo4jQueryParameter::add_value(int new_val) {
-  std::string err_msg;
-  if (!is_list) {
-    err_msg = "Trying to add value to non-arry parameter";
-    throw Neo4jException(err_msg);
-  } else {
-    if (type == -1 || type == _INT_TYPE) {
-      int_values.push_back(new_val);
-      type = _INT_TYPE;
-    } else {
-      err_msg = "Only single-type arrays are supported";
-      throw Neo4jException(err_msg);
-    }
-  }
-}
-
-void Neo4jQueryParameter::add_value(float new_val) {
-  if (!is_list) {
-    std::string err_msg = "Trying to add value to non-arry parameter";
-    throw Neo4jException(err_msg);
-  } else {
-    if (type == -1 || type == _FLT_TYPE) {
-      double_values.push_back(new_val);
-      type = _FLT_TYPE;
-    } else {
-      std::string err_msg = "Only single-type arrays are supported";
-      throw Neo4jException(err_msg);
-    }
-  }
-}
-
-unsigned int Neo4jQueryParameter::size() {
-  if (!is_list) {
-    std::string err_msg = "Trying to take the size of a non-arry parameter";
-    throw Neo4jException(err_msg);
-  } else {
-    if (type == _FLT_TYPE) {
-      return double_values.size();
-    } else if (type == _INT_TYPE) {
-      return int_values.size();
-    } else if (type == _STR_TYPE) {
-      return str_values.size();
-    } else if (type == _BOOL_TYPE) {
-      return bool_values.size();
-    } else {
-      return 0;
-    }
-  }
-}
-
-  // Neo4j Admin
-
   // Start the admin
-  void Neo4jAdmin::initialize(const char * conn_str, bool secure, int pool_size) {
+void Neo4jAdmin::initialize(const char * conn_str, bool secure, int pool_size) {
   // Initialize the Neo4j Database pool
   pool = new Neo4jConnectionPool(pool_size, conn_str, secure, 1, 1);
 }
@@ -629,175 +130,4 @@ ResultsIteratorInterface* Neo4jAdmin::execute(const char * query, \
 
   // Return a results iterator for results access
   return new ResultsIterator(res, qs, pool);
-}
-
-// Connection Pool
-
-void Neo4jConnectionPool::init_slots() {
-  slot_pool = new AOSSL::SlotPool(connection_limit);
-}
-
-// Initialize the connections required to start the connection pool
-void Neo4jConnectionPool::init_connections(const char * conn_str, bool secure) {
-  for (int i = 0; i < start_connections; i++) {
-    Neo4jQuerySession qs;
-
-    // Create the connection
-    if (!secure) {
-      qs.connection = neo4j_connect(conn_str, NULL, NEO4J_INSECURE);
-    } else {
-      qs.connection = neo4j_connect(conn_str, NULL, NEO4J_CONNECT_DEFAULT);
-    }
-
-    // Check if the connection was successful
-    if (!(qs.connection)) {
-      const char * err_string = strerror(errno);
-      std::string err_msg(err_string);
-      err_msg = "Error establishing connection: " + err_msg;
-      throw Neo4jException(err_msg);
-    }
-
-    // Establish a new session with the current connection
-    qs.session = neo4j_new_session(qs.connection);
-
-    // Check if session creation was successful
-    if (!(qs.session)) {
-      const char * err_string = strerror(errno);
-      std::string err_msg(err_string);
-      err_msg = "Error starting session: " + err_msg;
-      throw Neo4jException(err_msg);
-    }
-
-    connections.push_back(qs);
-  }
-  current_max_connection = start_connections;
-}
-
-Neo4jConnectionPool::~Neo4jConnectionPool() {
-  for (int i = 0; i < current_max_connection; i++) {
-    // Release the session
-    neo4j_end_session(connections[i].session);
-    // Release the connection
-    neo4j_close(connections[i].connection);
-  }
-  neo4j_client_cleanup();
-  delete slot_pool;
-}
-
-Neo4jQuerySession* Neo4jConnectionPool::get_connection() {
-  bool current_connection_existed = true;
-
-  // Find the next available connection slot
-  // If none are available, wait until one is freed
-  int current_connection = slot_pool->find_next_slot();
-
-  // Create new connections if necessary
-  if (current_connection > current_max_connection) {
-    // Get the creation mutex to ensure that only 1 thread creates new connections
-    std::lock_guard<std::mutex> lock(create_conn_mutex);
-    // Double check and see if any other threads built connections since we asked for the mutex
-    if (current_connection > current_max_connection) {
-      current_connection_existed = false;
-      for (
-        int i = current_max_connection;
-        i < current_max_connection + connection_creation_batch;
-        i++
-      ) {
-        Neo4jQuerySession qs;
-        if (!secure) {
-          qs.connection = neo4j_connect(connection_string.c_str(), \
-            NULL, NEO4J_INSECURE);
-        } else {
-          qs.connection = neo4j_connect(connection_string.c_str(), \
-            NULL, NEO4J_CONNECT_DEFAULT);
-        }
-
-        // Check if the connection was successful
-        if (!(qs.connection)) {
-          const char * err_string = strerror(errno);
-          std::string err_msg(err_string);
-          err_msg = "Error establishing connection: " + err_msg;
-          throw Neo4jException(err_msg);
-        }
-
-        // Establish a new session with the current connection
-        qs.session = neo4j_new_session(qs.connection);
-
-        // Check if session creation was successful
-        if (!(qs.session)) {
-          const char * err_string = strerror(errno);
-          std::string err_msg(err_string);
-          err_msg = "Error starting session: " + err_msg;
-          throw Neo4jException(err_msg);
-        }
-
-        connections.push_back(qs);
-      }
-      current_max_connection = current_max_connection+connection_creation_batch;
-    }
-  }
-
-  // Pack the connection & session into the return object
-  Neo4jQuerySession *s = new Neo4jQuerySession;
-  s->connection = connections[current_connection].connection;
-  s->session = connections[current_connection].session;
-  s->index = current_connection;
-
-  // Reset the Neo4j Session if it previously existed
-  if (current_connection_existed) {
-    if (neo4j_reset_session(s->session) != 0) {
-      const char * err_string = strerror(errno);
-      std::string err_msg(err_string);
-      err_msg = "Error resetting session: " + err_msg;
-      throw Neo4jException(err_msg);
-    }
-  }
-
-  // Return the latest connection information
-  // When we leave the method, the mutex lock is released automatically
-  return s;
-}
-
-// Release a connection for re-use
-void Neo4jConnectionPool::release_connection(Neo4jQuerySession *conn) {
-  // Release the slot
-  slot_pool->release_slot(conn->index);
-  // Delete the query session pointer, but not it's contents
-  delete conn;
-}
-
-Neo4jConnectionPool::Neo4jConnectionPool(int size, const char * conn_str, \
-  bool secure_connect) {
-  connection_limit = size;
-  std::string con_str(conn_str);
-  connection_string = con_str;
-  secure = secure_connect;
-  init_slots();
-  neo4j_client_init();
-  init_connections(conn_str, secure);
-}
-
-Neo4jConnectionPool::Neo4jConnectionPool(int size, const char * conn_str, \
-  bool secure_connect, int start_conns) {
-  connection_limit = size;
-  std::string con_str(conn_str);
-  connection_string = con_str;
-  secure = secure_connect;
-  start_connections = start_conns;
-  init_slots();
-  neo4j_client_init();
-  init_connections(conn_str, secure);
-}
-
-Neo4jConnectionPool::Neo4jConnectionPool(int size, const char * conn_str, \
-  bool secure_connect, int start_conns, int batch_size) {
-  connection_limit = size;
-  std::string con_str(conn_str);
-  connection_string = con_str;
-  secure = secure_connect;
-  start_connections = start_conns;
-  connection_creation_batch = batch_size;
-  init_slots();
-  neo4j_client_init();
-  init_connections(conn_str, secure);
 }
